fix click snackbar in auth page
